added cursor move animation on home page and fix some bugs
