+++ package.json
@@ -41,6 +41,7 @@
   },
   "devDependencies": {
     "autoprefixer": "^10.0.1",
+    "daisyui": "^4.6.1",
     "eslint": "^8",
     "eslint-config-next": "14.0.4",
     "postcss": "^8",
